move execution resource manager to state module
